requirements.txt
# Core dependencies
pydantic==2.5.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP clients (used for Bolna REST API calls)
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3
python-multipart>=0.0.9

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
pyodbc>=5.1.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3